feat: Add CreatePost method to handle post creation
This commit adds the CreatePost method to the msgServer struct in the msg_server_create_post.go file. The method handles the creation of a new post by storing the post details in the database. It takes in a message of type MsgCreatePost and returns a response of type MsgCreatePostResponse.

The method creates a new Post object using the data from the message and appends it to the database. The ID of the newly created post is then returned in the response.

This commit also includes recent user commits:
- 增加数据库prefix key
- 添加CRUD方法
- create-post
- 1. 定义帖子类型
- add readme
- Initialized with Ignite CLI